In the middle of adding main data structures and variables.
Added the PQ array. Size is PQMAXSIZE + 1 to count for the 0th element
not being used.
Added the Customer_t type structure and the comments to explain the
variables.
Added the n, lambda, mu, and M variables all "globally" defined in
Analytical.h.
Added the set function prototypes for all the variables above.